Refactored pages products and productForm to be children of pages/products folder as index and productForm respectivly
