+++ next-front/components/products/productList.jsx
@@ -12,7 +12,7 @@ export default function ProductList({ styles, products, setProducts}){
 
   return(
     <>
-    <div><h4>Total Number of Products: { parsedProducts.length }</h4></div>
+    
     <section className={ styles.list }>
       { parsedProducts }
     </section>
